revised draft of initial Wam API
- removing bank/preset methods and members
- consolidating onAutomation & onMidi into onEvent
- sketching out methods for managing event callbacks
- making WamProcessor implement getCompensationDelay
- moving type definitions to single file
- linting
